scripts/daily_sales_to_sheets/requirements.txt
# 日次売上 → Google Sheets 自動投入
python-dotenv>=1.0.0
requests>=2.28.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
